fix rendering below 0 and above 255 and for waterlogged top blocks

+++ src/main/java/net/querz/mcaselector/version/anvil112/Anvil112ChunkRenderer.java
@@ -11,7 +11,7 @@ import static net.querz.mcaselector.validation.ValidationHelper.withDefault;
 public class Anvil112ChunkRenderer implements ChunkRenderer {
 
 	@Override
-	public void drawChunk(CompoundTag root, ColorMapping colorMapping, int x, int z, int[] pixelBuffer, int[] waterPixels, byte[] terrainHeights, byte[] waterHeights, boolean water) {
+	public void drawChunk(CompoundTag root, ColorMapping colorMapping, int x, int z, int[] pixelBuffer, int[] waterPixels, short[] terrainHeights, short[] waterHeights, boolean water) {
 		ListTag<CompoundTag> sections = withDefault(() -> root.getCompoundTag("Level").getListTag("Sections").asCompoundTagList(), null);
 		if (sections == null) {
 			return;
@@ -65,7 +65,7 @@ public class Anvil112ChunkRenderer implements ChunkRenderer {
 							if (water) {
 								if (!waterDepth) {
 									pixelBuffer[regionIndex] = colorMapping.getRGB(((block << 4) + blockData)) | 0xFF000000;
-									waterHeights[regionIndex] = (byte) (sectionHeight + cy);
+									waterHeights[regionIndex] = (short) (sectionHeight + cy);
 								}
 								if (isWater(block)) {
 									waterDepth = true;
@@ -76,7 +76,7 @@ public class Anvil112ChunkRenderer implements ChunkRenderer {
 							} else {
 								pixelBuffer[regionIndex] = colorMapping.getRGB(((block << 4) + blockData)) | 0xFF000000;
 							}
-							terrainHeights[regionIndex] = (byte) (sectionHeight + cy);
+							terrainHeights[regionIndex] = (short) (sectionHeight + cy);
 							continue zLoop;
 						}
 					}
